requirements.txt
streamlit>=1.30.0
langchain>=0.1.0
langchain-community>=0.1.0
langchain-core>=0.1.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
huggingface-hub>=0.20.0
pyttsx3==2.90
comtypes>=1.4.11
pypiwin32>=223
pywin32>=310
requests>=2.31.0
numpy>=1.24.0
sentence-transformers>=2.2.2
pandas>=2.0.0
